+++ xtool/__init__.py
@@ -1,3 +1,3 @@
 """xtool - clean up your X (Twitter) account."""
 
-__version__ = "0.2.4"
+__version__ = "0.2.5"